[move] stylesファイルの移動
